fix: use type int for CampaignId (#86)

+++ src/Helldivers-2-Models/V1/Planets/Event.cs
@@ -20,6 +20,6 @@ public sealed record Event(
     long MaxHealth,
     DateTime StartTime,
     DateTime EndTime,
-    long CampaignId,
+    int CampaignId,
     List<int> JointOperationIds
 );
